functionapp/requirements.txt
azure-data-tables==12.0.0